requirements.txt
python-igraph==0.8.2
numpy==1.17.3
scipy==1.2.1
logging==0.5.1.2